refactor: Isolate show startup progress
